Warning: Dark magic incoming.
This commit introduces the concept of a morpher.

Morphers are things that can become any other given a strategy for generating
that type, and are simplified as whatever type they last became.

It is unlikely that anyone will use these directly. The plan is to have
them be the new implementation strategy for flatmap and stateful testing.