New implementation strategy for sets
Idea: We track how large a set is supposed to be, and figure out
which of the templates actually produce unique results at the time
we come to reify it.

When shrinking a set we might end up with an accidentally smaller
set because we introduced duplicates. We eliminate that using assume
rather than attempting to patch it up because that could cause a
continuous shrink/grow loop.